py/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
tldextract>=3.4.0
python-whois>=0.8.0
joblib>=1.2.0
streamlit>=1.28.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
